refactor Mac platform into POSIX base
For ease of comparsion with the original Mac-only source files,
we retain the line endings on the refactored POSIX common files to
exactly match the original ones.
