+++ GVFS/GVFS.Hooks/HooksPlatform/GVFSHooksPlatform.Mac.cs
@@ -1,4 +1,4 @@
-using GVFS.Platform.Mac;
+using GVFS.Platform.POSIX;
 
 namespace GVFS.Hooks.HooksPlatform
 {
@@ -6,37 +6,37 @@ namespace GVFS.Hooks.HooksPlatform
     {
         public static string GetInstallerExtension()
         {
-            return MacPlatform.InstallerExtension;
+            return "dmg";
         }
 
         public static bool IsElevated()
         {
-            return MacPlatform.IsElevatedImplementation();
+            return POSIXPlatform.IsElevatedImplementation();
         }
 
         public static bool IsProcessActive(int processId)
         {
-            return MacPlatform.IsProcessActiveImplementation(processId);
+            return POSIXPlatform.IsProcessActiveImplementation(processId);
         }
 
         public static string GetNamedPipeName(string enlistmentRoot)
         {
-            return MacPlatform.GetNamedPipeNameImplementation(enlistmentRoot);
+            return POSIXPlatform.GetNamedPipeNameImplementation(enlistmentRoot);
         }
 
         public static bool IsConsoleOutputRedirectedToFile()
         {
-            return MacPlatform.IsConsoleOutputRedirectedToFileImplementation();
+            return POSIXPlatform.IsConsoleOutputRedirectedToFileImplementation();
         }
 
         public static bool TryGetGVFSEnlistmentRoot(string directory, out string enlistmentRoot, out string errorMessage)
         {
-            return MacPlatform.TryGetGVFSEnlistmentRootImplementation(directory, out enlistmentRoot, out errorMessage);
+            return POSIXPlatform.TryGetGVFSEnlistmentRootImplementation(directory, out enlistmentRoot, out errorMessage);
         }
 
         public static bool TryGetNormalizedPath(string path, out string normalizedPath, out string errorMessage)
         {
-            return MacFileSystem.TryGetNormalizedPathImplementation(path, out normalizedPath, out errorMessage);
+            return POSIXFileSystem.TryGetNormalizedPathImplementation(path, out normalizedPath, out errorMessage);
         }
     }
 }
