+++ GVFS/GVFS.Platform.POSIX/POSIXFileSystem.Shared.cs
@@ -1,10 +1,10 @@
-﻿namespace GVFS.Platform.Mac
+﻿namespace GVFS.Platform.POSIX
 {
-    public partial class MacFileSystem
+    public partial class POSIXFileSystem
     {
         public static bool TryGetNormalizedPathImplementation(string path, out string normalizedPath, out string errorMessage)
         {
-            // TODO(Mac): Properly determine normalized paths (e.g. across links)
+            // TODO(POSIX): Properly determine normalized paths (e.g. across links)
             errorMessage = null;
             normalizedPath = path;
             return true;
